chore(hostnames): update generated types and methods

+++ src/resources/hostnames/settings/tls.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../resource';
+import { isRequestOptions } from '../../../core';
 import * as Core from '../../../core';
 import { SinglePage } from '../../../pagination';
 
@@ -14,7 +15,7 @@ export class TLS extends APIResource {
     params: TLSUpdateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<Setting> {
-    const { zone_id, ...body } = params;
+    const { zone_id = this._client.zoneId, ...body } = params;
     return (
       this._client.put(`/zones/${zone_id}/hostnames/settings/${settingId}/${hostname}`, {
         body,
@@ -29,10 +30,24 @@ export class TLS extends APIResource {
   delete(
     settingId: 'ciphers' | 'min_tls_version' | 'http2',
     hostname: string,
-    params: TLSDeleteParams,
+    params?: TLSDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<TLSDeleteResponse>;
+  delete(
+    settingId: 'ciphers' | 'min_tls_version' | 'http2',
+    hostname: string,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<TLSDeleteResponse>;
+  delete(
+    settingId: 'ciphers' | 'min_tls_version' | 'http2',
+    hostname: string,
+    params: TLSDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<TLSDeleteResponse> {
-    const { zone_id } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(settingId, hostname, {}, params);
+    }
+    const { zone_id = this._client.zoneId } = params;
     return (
       this._client.delete(
         `/zones/${zone_id}/hostnames/settings/${settingId}/${hostname}`,
@@ -46,10 +61,22 @@ export class TLS extends APIResource {
    */
   get(
     settingId: 'ciphers' | 'min_tls_version' | 'http2',
-    params: TLSGetParams,
+    params?: TLSGetParams,
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<TLSGetResponsesSinglePage, TLSGetResponse>;
+  get(
+    settingId: 'ciphers' | 'min_tls_version' | 'http2',
+    options?: Core.RequestOptions,
+  ): Core.PagePromise<TLSGetResponsesSinglePage, TLSGetResponse>;
+  get(
+    settingId: 'ciphers' | 'min_tls_version' | 'http2',
+    params: TLSGetParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.PagePromise<TLSGetResponsesSinglePage, TLSGetResponse> {
-    const { zone_id } = params;
+    if (isRequestOptions(params)) {
+      return this.get(settingId, {}, params);
+    }
+    const { zone_id = this._client.zoneId } = params;
     return this._client.getAPIList(
       `/zones/${zone_id}/hostnames/settings/${settingId}`,
       TLSGetResponsesSinglePage,
@@ -195,7 +222,7 @@ export interface TLSUpdateParams {
   /**
    * Path param: Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 
   /**
    * Body param: The TLS setting value. The type depends on the `setting_id` used in
@@ -215,14 +242,14 @@ export interface TLSDeleteParams {
   /**
    * Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 }
 
 export interface TLSGetParams {
   /**
    * Identifier.
    */
-  zone_id: string;
+  zone_id?: string;
 }
 
 TLS.TLSGetResponsesSinglePage = TLSGetResponsesSinglePage;
